feat: shppingAPI 별도 생성  (#204)
