+++ itda-front/src/util/API/cartAPI.ts
@@ -1,4 +1,4 @@
-import { ICartProduct,IShippingInfos, IOrder } from "types/CartTypes";
+import { ICartProduct, IShippingInfos, IOrder } from "types/CartTypes";
 import { instanceWithAuth } from "./index";
 
 const getCartProductList = () => instanceWithAuth.get("/cart");
@@ -9,25 +9,15 @@ const updateCartProductList = (productList: ICartProduct[]) =>
 const deleteCartProduct = (productId: string) =>
   instanceWithAuth.delete(`/cart?productId=${productId}`);
 
-const postShippingInfos = async (shippingInfos: IShippingInfos) => {
-  instanceWithAuth.post("/shippingInfos", shippingInfos);
-};
-
-const getShippingInfos = () => instanceWithAuth.get("/shippingInfos");
-
 const postOrder = async (orderList: IOrder) => {
   instanceWithAuth.post("/order", orderList);
 };
-
 
 const cartAPI = {
   get: { getCartProductList },
   post: { updateCartProductList },
   delete: { deleteCartProduct },
-   shipping: {
-    post: { postShippingInfos },
-    get: { getShippingInfos },
-  },
+
   order: {
     post: { postOrder },
   },
