+++ itda-front/src/util/API/shippingAPI.ts
@@ -1,1 +1,14 @@
-export {};
+import { ICartProduct, IShippingInfos, IOrder } from "types/CartTypes";
+import { instanceWithAuth } from "./index";
+
+const postShippingInfos = async (shippingInfos: IShippingInfos) => {
+  instanceWithAuth.post("/shippingInfos", shippingInfos);
+};
+
+const getShippingInfos = () => instanceWithAuth.get("/shippingInfos");
+
+const shippingAPI = {
+  post: { postShippingInfos },
+  get: { getShippingInfos },
+};
+export default shippingAPI;
